Modify blockchain tracker to handle users notifications

+++ engrave-api-gateway/src/routes/blogs/routes/removeBlog.ts
@@ -13,6 +13,11 @@ async function handler(req: Request, res: Response) {
     return handleResponseError(async () => {
         
         const { id } = req.body;
+        const { username } = res.locals;
+
+        const blog = await blogsService.getBlogByQuery({_id: id});
+        
+        if(blog.username != username) throw new Error('You are not the owner of that blog!');
         
         await blogsService.removeBlogWithQuery({_id: id});
